+++ packages/fxa-settings/src/lib/auth-errors/auth-errors.ts
@@ -3,10 +3,6 @@
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
 import { AuthServerError } from 'fxa-auth-client/browser';
-import * as Sentry from '@sentry/browser';
-import { FtlMsgResolver } from 'fxa-react/lib/utils';
-import { HandledError } from '../interfaces';
-import { OAUTH_ERRORS } from './../oauth/oauth-errors';
 
 export type AuthUiError = AuthServerError & { version?: number };
 
@@ -632,97 +628,6 @@ const ERRORS = {
 
 type ErrorKey = keyof typeof ERRORS;
 type ErrorVal = { errno: number; message: string; version?: number };
-
-/**
- * Utility function to retrieve the localized auth client error message
- * - works for throttling errors that include a localized retry after value
- * - returns an unexpected error if the error is unknown or does not have a localized message
- * @param ftlMsgResolver
- * @param err is an AuthClient error
- * @returns
- */
-
-export const getLocalizedErrorMessage = (
-  ftlMsgResolver: FtlMsgResolver,
-  error: AuthUiError | HandledError
-) => {
-  if (error.errno) {
-    if (AuthUiErrorNos[error.errno]) {
-      if (
-        error.retryAfterLocalized &&
-        error.errno === AuthUiErrors.THROTTLED.errno
-      ) {
-        // For throttling errors where a localized retry after value is provided
-        return ftlMsgResolver.getMsg(
-          getErrorFtlId(error),
-          AuthUiErrorNos[error.errno].message,
-          { retryAfter: error.retryAfterLocalized }
-        );
-      } else if (error.errno === AuthUiErrors.THROTTLED.errno) {
-        // For throttling errors where a localized retry after value is not available
-        return ftlMsgResolver.getMsg(
-          'auth-error-114-generic',
-          AuthUiErrorNos[114].message
-        );
-      } else {
-        // for all other recognized auth UI errors
-        return ftlMsgResolver.getMsg(
-          getErrorFtlId(error),
-          AuthUiErrorNos[error.errno].message
-        );
-      }
-    }
-    const oAuthError = Object.values(OAUTH_ERRORS).find(
-      (oAuthErr) => error.errno === oAuthErr.errno
-    );
-    if (oAuthError) {
-      return ftlMsgResolver.getMsg(
-        getErrorFtlId(oAuthError),
-        oAuthError.message
-      );
-    }
-  }
-
-  const unexpectedError = AuthUiErrors.UNEXPECTED_ERROR;
-  return ftlMsgResolver.getMsg(
-    getErrorFtlId(unexpectedError),
-    unexpectedError.message
-  );
-};
-
-/*
-  TODO in FXA-9502: account for potential errno overlap between auth-errors
-  and oauth errors. Checking the auth-error array currently happens first.
-**/
-export const getErrorFtlId = (err: { errno?: number; message?: string }) => {
-  if (err.errno) {
-    if (err.errno in AuthUiErrorNos) {
-      const error = AuthUiErrorNos[err.errno];
-      return `auth-error-${err.errno}${
-        error.version ? '-' + error.version : ''
-      }`;
-    }
-
-    const oAuthError = Object.values(OAUTH_ERRORS).find(
-      (oAuthErr) => err.errno === oAuthErr.errno
-    );
-    if (oAuthError) {
-      return `oauth-error-${err.errno}${
-        oAuthError.version ? '-' + oAuthError.version : ''
-      }`;
-    }
-  }
-  // If the error isn't found, return an empty string FTL ID and log to Sentry.
-  const logMessage = err.errno
-    ? `WARNING: An error occurred that we attempted to localize and render, but this error was not found in auth-errors or oauth-errors. We should either add this error to our list or not display it. error: ${JSON.stringify(
-        err
-      )}`
-    : `WARNING: An error occurred that we attempted to localize and render, but 'errno' is missing. error: ${JSON.stringify(
-        err
-      )}`;
-  Sentry.captureMessage(logMessage);
-  return '';
-};
 
 export const AuthUiErrors: { [key in ErrorKey]: AuthUiError } = (
   Object.entries(ERRORS) as [[ErrorKey, ErrorVal]]
