+++ packages/fxa-settings/src/pages/Signup/interfaces.ts
@@ -2,7 +2,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
-import { HandledError } from '../../lib/interfaces';
+import { HandledError } from '../../lib/error-utils';
 import {
   BaseIntegration,
   IntegrationType,
